rename normlize-event util
