+++ new-frontend/components/EventMonolog/EventMonolog.stories.ts
@@ -1,5 +1,5 @@
 import { Meta, Story } from "@storybook/vue3";
-import { normalizeMonologEvent } from "~/utils/normalizeEvent";
+import { normalizeMonologEvent } from "~/utils/normalize-event";
 import EventMonolog from '~/components/EventMonolog/EventMonolog.vue';
 import monologEventMock from '~/mocks/monolog.json'
