+++ new-frontend/utils/normalize-event.ts
